requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
setuptools==75.8.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
virtualenv==20.29.2
Werkzeug==3.1.3
